backend-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
qdrant-client>=1.9.0
cohere>=4.42
openai>=1.0.0
psycopg2-binary>=2.9.0
asyncpg>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sse-starlette>=1.8.2
python-multipart>=0.0.6